+++ app/Http/Controllers/Organization/OrganizationController.php
@@ -7,6 +7,7 @@ use App\Models\Organization;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Gate;
 
 class OrganizationController extends Controller
 {
@@ -22,9 +23,13 @@ class OrganizationController extends Controller
      */
     public function index()
     {
-        $organizations = Organization::all();
+        if (Gate::allows('is-manager-or-agent')) {
+            $organizations = Organization::all();
 
-        return response()->json($organizations);
+            return response()->json($organizations);
+        } else {
+            return response()->json(['message' => 'Access denied'], 403);
+        }
     }
 
     /**
@@ -35,16 +40,20 @@ class OrganizationController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'name' => 'required|string',
-            'phone_number' => 'required|string',
-            'description' => 'string',
-            'address' => 'required|string',
-        ]);
+        if (Gate::allows('is-manager-or-agent')) {
+            $request->validate([
+                'name' => 'required|string',
+                'phone_number' => 'required|string',
+                'description' => 'string',
+                'address' => 'required|string',
+            ]);
 
-        $organization = Organization::create($request->all());
+            $organization = Organization::create($request->all());
 
-        return response()->json($organization, 201);
+            return response()->json($organization, 201);
+        } else {
+            return response()->json(['message' => 'Access denied'], 403);
+        }
     }
 
     /**
@@ -55,9 +64,13 @@ class OrganizationController extends Controller
      */
     public function show($id)
     {
-        $organization = Organization::findOrFail($id);
+        if (Gate::allows('is-manager-or-agent')) {
+            $organization = Organization::findOrFail($id);
 
-        return response()->json($organization);
+            return response()->json($organization);
+        } else {
+            return response()->json(['message' => 'Access denied'], 403);
+        }
     }
 
     /**
@@ -69,17 +82,21 @@ class OrganizationController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $request->validate([
-            'name' => 'string',
-            'phone_number' => 'string',
-            'description' => 'string',
-            'address' => 'string',
-        ]);
+        if (Gate::allows('is-manager-or-agent')) {
+            $request->validate([
+                'name' => 'string',
+                'phone_number' => 'string',
+                'description' => 'string',
+                'address' => 'string',
+            ]);
 
-        $organization = Organization::findOrFail($id);
-        $organization->update($request->all());
+            $organization = Organization::findOrFail($id);
+            $organization->update($request->all());
 
-        return response()->json($organization);
+            return response()->json($organization);
+        } else {
+            return response()->json(['message' => 'Access denied'], 403);
+        }
     }
 
     /**
@@ -90,9 +107,13 @@ class OrganizationController extends Controller
      */
     public function destroy($id)
     {
-        $organization = Organization::findOrFail($id);
-        $organization->delete();
+        if (Gate::allows('is-manager-or-agent')) {
+            $organization = Organization::findOrFail($id);
+            $organization->delete();
 
-        return response()->json(null, 204);
+            return response()->json(null, 204);
+        } else {
+            return response()->json(['message' => 'Access denied'], 403);
+        }
     }
 }
